+++ trunk/src/tk/svg/svg_gradient.h
@@ -1,7 +1,7 @@
 // ======================================================================
 // Author : $Author$
-// Version: $Revision: 1 $
-// Date   : $Date: 2011-05-04 00:04:08 +0000 (Wed, 04 May 2011) $
+// Version: $Revision: 28 $
+// Date   : $Date: 2011-05-21 14:57:26 +0000 (Sat, 21 May 2011) $
 // Url    : $URL$
 // ======================================================================
 
@@ -74,7 +74,7 @@ public:
 
 	color_type operator[] (unsigned v) const
 	{
-		return m_color_list[mstl::min(v, m_color_list.size() - 1)];
+		return m_color_list[mstl::min(size_t(v), m_color_list.size() - 1)];
 	}
 
 private:
