+++ trunk/src/util/libTeXt/T_UserInput.cpp
@@ -1,7 +1,7 @@
 // ======================================================================
 // Author : $Author$
-// Version: $Revision: 1 $
-// Date   : $Date: 2011-05-04 00:04:08 +0000 (Wed, 04 May 2011) $
+// Version: $Revision: 28 $
+// Date   : $Date: 2011-05-21 14:57:26 +0000 (Sat, 21 May 2011) $
 // Url    : $URL$
 // ======================================================================
 
@@ -46,7 +46,7 @@ UserInput::readNextLine(mstl::string& result)
 	mstl::string::size_type n = m_text.find('\n');
 
 	result.assign(m_text, 0, n);
-	m_text.erase(0u, n);
+	m_text.erase(size_t(0), n);
 
 	return true;
 }
